+++ src/guards/isPersonSaved.ts
@@ -4,28 +4,7 @@ import useFieldsState from '../composables/useFieldsState'
 const { showDialog, dialogState } = useDialog();
 const { isPristineState, statesID } = useFieldsState();
 
-
-export interface DialogHeader {
-    title: string;
-}
-
-export interface DialogBody {
-    icon: string;
-    answer: string;
-    colorClass: string;
-    comment: string;
-}
-
-export interface DialogOptions {
-    nameRouteToRedirect?: string;
-    routeProp?: string;
-    header: DialogHeader;
-    body: DialogBody;
-    footer?: any;
-}
-
 const keyMap: { [key: string]: string } = {
-    // claveOriginal: 'Clave formateada'
     typeDocumento: 'Tipo de documento',
     fecha: 'Fecha de nacimiento',
     estadoCivil: 'Estado Civil',
@@ -37,25 +16,34 @@ const keyMap: { [key: string]: string } = {
 
 const formatKey = (key: string) => {
     return keyMap[key] || key;
-};;
+};
 
 const isPersonSaved = (to, from) => {
-
     const pathFindGuard = ['personas'];
     const pathIncludesGuard = pathFindGuard.some(keyword => from.path.includes(keyword));
+
     if (pathIncludesGuard && dialogState.value.pendingRoute === null) {
-
         if (pathIncludesGuard && isPristineState.value) {
-
             const modifiedItem = statesID.find(item => Object.keys(item.modifiedData).length > 0);
 
             if (modifiedItem) {
-                // Obtener las claves de modifiedData y formatearlas
                 const modifiedDataKeys = Object.keys(modifiedItem.modifiedData)
                     .map(formatKey)
                     .join(', ');
-                alert(modifiedDataKeys)
-                return true;
+
+                showDialog({
+                    nameRouteToRedirect: to.path,
+                    routeProp: 'path',
+                    header: { title: 'Cambios sin guardar' },
+                    body: {
+                        icon: 'warning',
+                        comment: `Hay cambios sin guardar en los siguientes campos: ${modifiedDataKeys}`,
+                        colorClass: 'warning',
+                        answer: '¿Desea salir sin guardar?'
+                    },
+                    footer: 'customFooterComponent'
+                });
+                return false;
             }
         }
     }
@@ -66,7 +54,7 @@ const isPersonSaved = (to, from) => {
         return true;
     }
 
-    return true
+    return true;
 };
 
 export default isPersonSaved;
