+++ src/interfaces/dialogInterfaces.ts
@@ -1,0 +1,18 @@
+export interface DialogHeader {
+    title: string;
+}
+
+export interface DialogBody {
+    icon: string;
+    answer: string;
+    colorClass: string;
+    comment: string;
+}
+
+export interface DialogOptions {
+    nameRouteToRedirect?: string;
+    routeProp?: string;
+    header: DialogHeader;
+    body: DialogBody;
+    footer?: any; //quizas lo usemos mas tarde
+}
